snap, systemd, wrappers: move RestartCondition from systemd to snap
